requirements.txt
# requirements.txt

streamlit>=1.30.0
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2